+++ CHANGELOG.md
@@ -10,10 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - In interactive plotlib, Support for 2d plot, better data manipulation and better fits
 - config builder MeasurePulse now accepts IntegrationWeights object as well
+- config builder Element now accepts ControlPulse and MeasurePulse
+- config builder Parameter class now support basic algebra (+, -, /, *, **)
 ### Changed
 - All config builder objects can be initialized with all the data (but still can be built step by step)
 - config builder FluxTunableTransmon - parameter name changed from fl_port to flux_port
 - config builder Coupler - parameter name changed from p to port
+- config builder Element - renamed set_delay/buffer methods as set_digital_input_delay/buffer
+- config builder ConfigVar is changed to ConfigVars
+- config builder ConfigVars - parameter method now returns a Parameter object instead of lambda function
+- config builder ConfigVars - parameter method optionally accepts a setter
 
 
 ## [0.8.0] - 2022-04-04